+++ src/kubernetes-tool/descriptions/Deployment.yml
