+++ src/kubernetes-tool/descriptions/StatefulSet.yml
