+++ src/kubernetes-tool/templates/properties.vue
@@ -26,8 +26,6 @@ limitations under the License.
 </template>
 
 <script>
-    import { flags } from "../utils/flags"
-
     export default {
         name: "Properties",
         props: {
@@ -38,7 +36,7 @@ limitations under the License.
             handleArrayTrimming(v) {
                 const copy = []
                 for (i of v) {
-                    if (flags.includes("show-undefined") || i.value) copy.push(i)
+                    if (i.value) copy.push(i)
                 }
                 return copy
             },
