--- src/kubernetes-tool/utils/flags.ts
@@ -1,4 +1,0 @@
-export const flags: string[] = []
-
-// @ts-ignore
-window.addFlag = (flag: string) => flags.push(flag)
